feat: refactor ads pages with shared progress and form components
Introduce reusable progress steps and multi-step form components.

#VERCEL_SKIP
